Refactor VolumetricClouds and clean up dead code
- Delete unused AsciiClouds component
- Extract ControlPanel into its own file
- Extract Controls interface, DEFAULTS, and SKY_PRESETS into controls.ts
- Remove unused rgbToHex function and no-op "use client" directive
- Fix scroll-driven values to derive from DEFAULTS instead of hardcoded numbers
- Update flight-tracker card gradient to match oatmeal/bark palette
