+++ pages/projects/flight-tracker/flight-tracker.css
@@ -1,8 +1,8 @@
 .flightTracker__background {
   background: radial-gradient(
     77.37% 50% at 50% 100%,
-    #1a3a5c 0%,
-    #0f2440 72.87%,
-    #091728 100%
+    #f4d6b0 0%,
+    #c9a87a 72.87%,
+    #341109 100%
   );
 }
