+++ pages/projects/flight-tracker/index.tsx
@@ -1,5 +1,3 @@
-"use client";
-
 import { useEffect, useState } from "react";
 import dynamic from "next/dynamic";
 import Content from "@/components/Layout/Content";
